+++ Apply-For-Me-Api/src/main/java/com/hydraulic/applyforme/service/JobSubmissionService.java
@@ -1,5 +1,5 @@
 package com.hydraulic.applyforme.service;
 
 public interface JobSubmissionService {
-    public Integer countAllSubmissions(Long Id);
+    public Long countAllSubmissions(Long Id);
 }
